Extend members, JSON, and formatting to big integers under sandbox guards
Every int member now handles big receivers: abs/succ/pred promote at the
int64 boundary instead of erroring, even?/odd? and the sign predicates
consult the payload, round/floor/ceil bucket in arbitrary precision
(Ruby-checked, including float rounding beyond int64 and Numeric#div/
divmod/modulo/remainder), clamp/between? compare through big.Rat, and
to_f/fdiv/Math/%float formatting convert best-effort. The iteration
members (times/upto/downto/step, range.step) and the int64 domains
(Duration units, Time constructors and Time.at, sleep, rand/srand,
counts, indexes, precisions, task limits) reject big values loudly —
never a silent Value.Int truncation to zero. JSON now parses integer
tokens beyond int64 as big integers (previously degraded to floats) and
stringifies them as bare decimals; %d/%b/%o/%O/%x format big operands
natively with size-aware projections; range.sum spills into an
arbitrary-precision accumulator instead of raising.

Sandbox guards (the #604/#913 conventions): big-operand arithmetic
charges steps proportional to word count (1 + words/8, via a new stepN),
exponentiation preflights exp x log2(base) bits against both quotas in
O(1) so 2 ** 10_000_000_000 rejects instantly, oversized products and
10^digits rounding buckets preflight before allocating, and every
rendering path (to_s, inspect, interpolation, puts, JSON, format)
charges digit-scaled steps and projected bytes before the superlinear
base conversion runs. Big payloads dedupe through the estimator's
existing slice-identity space, keeping the journal structs at their
pre-bignum sizes so the memoized walk path stays inlined (arithmetic
loop stays within ~3% of master with identical allocations).

Existing tests updated to pin the new semantics (each previously pinned
a replaced int64-overflow error or the JSON float degradation):
TestIntegerSuccessorOverflow, TestNumericDivisionOverflow,
TestNumericRoundingOverflow, TestFloatNegativePrecisionBucketsBeyondInt64,
TestFloatNegativePrecisionExtremeDigits, TestMethodErrorHandling (moved
to TestFloatRoundingBeyondInt64Promotes), TestRangeSumOverflow,
TestJSONParseEscapesAndNumbers, and TestRoundMinIntPrecisionDoesNotOverflow
(bigIntRound became the promoting bigIntRoundValue). Fuzz corpora seeded
with big literals. Fixture A/B against master: 74 fixtures byte-identical.

+++ internal/runtime/math.go
@@ -151,7 +151,9 @@ func builtinMathLog(_ *Execution, _ Value, args []Value, kwargs map[string]Value
 func mathFloatArg(name string, arg Value) (float64, error) {
 	switch arg.Kind() {
 	case KindInt:
-		return float64(arg.Int()), nil
+		// Value.Float converts big integers best-effort (saturating to the
+		// infinities), identical to float64(Int()) for compact values.
+		return arg.Float(), nil
 	case KindFloat:
 		return arg.Float(), nil
 	default:
